Built prototype requirements method for finding valid plays. It needs a lot of work but it's a great base.
"Built" should be "built and rebuilt and rebuilt.."

I know it's not much, but this commit is dedicated to Jared Harless. What could push such a good man to do such a terrible thing is beyond me. No one should have to feel the pain he felt. May he and his family rest in piece.